Simulate data for minimisation simulations

Given a list of factors with specified levels and proportions, and a sample size, this function simulates data for use in simulate_mini().

Usage

simulate_data(sampsize, factors, eligibility = FALSE, groups = NULL)

Arguments

sampsize

the desired sample size of the data, i.e., the sample size of your prospective trial.

factors

a list of factors, each a list containing two items. The first is levels which is either a vector of the level names OR the number of levels. The second item is either props, a vector of proportions equal to the number of levels in the factor; OR prop.dist which is a named vector containing the mean and sd of the proportions.

eligibility

logical, should the data include group eligibility. If TRUE then must supply groups argument.

groups

either a character vector of group names or an integer specifying the number of groups.

Examples

simulate_data(factors = list(sex = list(levels = c("M", "F"),
                                        props = c(0.5, 0.5)),
                             site = list(levels = 12,
                                         prop.dist = c(mean = 0.1, sd = 0.05))),
              sampsize = 150)
#> Error in sample.int(length(x), size, replace, prob): invalid first argument

simulate_data(factors = list(sex = list(levels = c("M", "F"),
                                        props = c(0.5, 0.5)),
                             site = list(levels = 12,
                                         prop.dist = c(mean = 0.1, sd = 0.05))),
              sampsize = 100, eligibility = TRUE, groups = 4)
#>      ID sex site eligible
#> 1     1   M    5       CD
#> 2     2   M    7      ACD
#> 3     3   F    6       AC
#> 4     4   F   10       CD
#> 5     5   F    8      ABD
#> 6     6   F   11       BD
#> 7     7   F    5       AD
#> 8     8   M   12      BCD
#> 9     9   F    5     ABCD
#> 10   10   F    8      ACD
#> 11   11   F    5      ABC
#> 12   12   F    1       BD
#> 13   13   M   10      ABD
#> 14   14   M    2      BCD
#> 15   15   M   11       BC
#> 16   16   M    9      ABC
#> 17   17   F    1     ABCD
#> 18   18   M    9       BC
#> 19   19   F   10      BCD
#> 20   20   M    8       CD
#> 21   21   F    3       AB
#> 22   22   M    1       CD
#> 23   23   F   10       CD
#> 24   24   M   12     ABCD
#> 25   25   F    5       AC
#> 26   26   F    3       CD
#> 27   27   F    6      ABD
#> 28   28   M    9     ABCD
#> 29   29   F    2       CD
#> 30   30   F   12       AD
#> 31   31   F    4      ABC
#> 32   32   M   12      ABD
#> 33   33   M   12      ABC
#> 34   34   M    8       BD
#> 35   35   F    9       CD
#> 36   36   M    8       BC
#> 37   37   F    5       BC
#> 38   38   F    8      ACD
#> 39   39   M   12       BC
#> 40   40   F    8     ABCD
#> 41   41   M    5     ABCD
#> 42   42   F    6      ACD
#> 43   43   M   11      ACD
#> 44   44   F   10      ABD
#> 45   45   M   10     ABCD
#> 46   46   M    6       BC
#> 47   47   F   10      ABD
#> 48   48   F    4       AC
#> 49   49   M    5      ABC
#> 50   50   F    5       BD
#> 51   51   M    5       BC
#> 52   52   F    8      ABC
#> 53   53   M    5       BC
#> 54   54   F    6      ABC
#> 55   55   F    4       AB
#> 56   56   M    9       AB
#> 57   57   M    1      ABD
#> 58   58   M   11     ABCD
#> 59   59   F    1       AD
#> 60   60   F    3      BCD
#> 61   61   F    5      ACD
#> 62   62   M    6       AD
#> 63   63   F    6      ABC
#> 64   64   M    8      ABC
#> 65   65   F   11       BD
#> 66   66   M    4       CD
#> 67   67   F    4      ABD
#> 68   68   M   12      ABC
#> 69   69   F    5       CD
#> 70   70   F    8      BCD
#> 71   71   M   10      ABD
#> 72   72   M    4       AC
#> 73   73   F   12       AC
#> 74   74   F   12       BD
#> 75   75   M    8       BD
#> 76   76   M    4      BCD
#> 77   77   F    4       AD
#> 78   78   F    1       BD
#> 79   79   M    8     ABCD
#> 80   80   M    1     ABCD
#> 81   81   M    5       AC
#> 82   82   F    5       BC
#> 83   83   M    5      ACD
#> 84   84   F    1     ABCD
#> 85   85   F    5      ACD
#> 86   86   F    8       BC
#> 87   87   F   10       BC
#> 88   88   M   11      ABD
#> 89   89   M   11       AD
#> 90   90   M    2       BD
#> 91   91   M    5      ABC
#> 92   92   F   11       BD
#> 93   93   M    8      BCD
#> 94   94   F    5       CD
#> 95   95   M    5      ABC
#> 96   96   M    2     ABCD
#> 97   97   F    5      BCD
#> 98   98   M    4      BCD
#> 99   99   M    3       AD
#> 100 100   M    2       BC