Simulate data for minimisation simulations

Given a list of factors with specified levels and proportions, and a sample size, this function simulates data for use in simulate_mini().

Usage

simulate_data(sampsize, factors)

Arguments

sampsize

the desired sample size of the data, i.e., the sample size of your prospective trial.

factors

a list of factors, each a list containing two items. The first is levels which is either a vector of the level names OR the number of levels. The second item is either props, a vector of proportions equal to the number of levels in the factor; OR prop.dist which is a named vector containing the mean and sd of the proportions.

Examples

simulate_data(factors = list(sex = list(levels = c("M", "F"),
                                        props = c(0.5, 0.5)),
                             site = list(levels = 12,
                                         prop.dist = c(mean = 0.1, sd = 0.05))),
              sampsize = 150)
#>      ID sex site
#> 1     1   M   12
#> 2     2   M    1
#> 3     3   M    2
#> 4     4   F    5
#> 5     5   F   10
#> 6     6   F   10
#> 7     7   M    3
#> 8     8   F   12
#> 9     9   F    4
#> 10   10   F    3
#> 11   11   M    5
#> 12   12   F   12
#> 13   13   F    4
#> 14   14   M    3
#> 15   15   M    9
#> 16   16   F    4
#> 17   17   F   12
#> 18   18   M    6
#> 19   19   M   11
#> 20   20   F    3
#> 21   21   F    3
#> 22   22   M    5
#> 23   23   F    1
#> 24   24   M    6
#> 25   25   M    9
#> 26   26   F    3
#> 27   27   M    4
#> 28   28   M    6
#> 29   29   F    5
#> 30   30   F    8
#> 31   31   F    8
#> 32   32   M    8
#> 33   33   M    3
#> 34   34   F    5
#> 35   35   M    2
#> 36   36   F    4
#> 37   37   M   10
#> 38   38   F    9
#> 39   39   F    4
#> 40   40   M   10
#> 41   41   M    2
#> 42   42   M    5
#> 43   43   M    3
#> 44   44   M   12
#> 45   45   F    3
#> 46   46   M   12
#> 47   47   M   12
#> 48   48   F    6
#> 49   49   M    8
#> 50   50   F    7
#> 51   51   F   10
#> 52   52   F   12
#> 53   53   F   11
#> 54   54   M   10
#> 55   55   F   12
#> 56   56   M    6
#> 57   57   F    3
#> 58   58   F   10
#> 59   59   M    3
#> 60   60   M    5
#> 61   61   F    7
#> 62   62   F    1
#> 63   63   M   12
#> 64   64   F   12
#> 65   65   F   10
#> 66   66   F    4
#> 67   67   F   12
#> 68   68   M    2
#> 69   69   M    5
#> 70   70   F    3
#> 71   71   M    6
#> 72   72   M    1
#> 73   73   F   12
#> 74   74   F    1
#> 75   75   F    1
#> 76   76   F    3
#> 77   77   F    7
#> 78   78   F    6
#> 79   79   F    9
#> 80   80   F   12
#> 81   81   F    9
#> 82   82   M   10
#> 83   83   M   10
#> 84   84   M   12
#> 85   85   F    7
#> 86   86   F    6
#> 87   87   F    6
#> 88   88   M    8
#> 89   89   M    6
#> 90   90   M    8
#> 91   91   M   10
#> 92   92   F    6
#> 93   93   F    8
#> 94   94   M    3
#> 95   95   F    5
#> 96   96   M   10
#> 97   97   M    7
#> 98   98   F    3
#> 99   99   M   12
#> 100 100   F   12
#> 101 101   M    8
#> 102 102   M    3
#> 103 103   M    3
#> 104 104   M    5
#> 105 105   M    6
#> 106 106   M    3
#> 107 107   M    2
#> 108 108   M    3
#> 109 109   F   10
#> 110 110   F    7
#> 111 111   M    4
#> 112 112   F    5
#> 113 113   F    6
#> 114 114   M    4
#> 115 115   M    3
#> 116 116   F   12
#> 117 117   M    3
#> 118 118   M    3
#> 119 119   M   10
#> 120 120   M    9
#> 121 121   M    2
#> 122 122   F    6
#> 123 123   F    1
#> 124 124   F    3
#> 125 125   M   10
#> 126 126   M    4
#> 127 127   M   12
#> 128 128   M    1
#> 129 129   F    2
#> 130 130   F    3
#> 131 131   F    4
#> 132 132   M    8
#> 133 133   F   10
#> 134 134   M    4
#> 135 135   F   12
#> 136 136   F    3
#> 137 137   M   12
#> 138 138   M    3
#> 139 139   M    4
#> 140 140   M    3
#> 141 141   F   12
#> 142 142   M    3
#> 143 143   F   12
#> 144 144   F    3
#> 145 145   M    8
#> 146 146   M    6
#> 147 147   M    6
#> 148 148   F   10
#> 149 149   M    4
#> 150 150   F    7